acquisition: search grids live only while searching — peak RSS −39%
sum_p was allocated full-width (75 bins × code_sp × f64 = 30 MB/channel
at 50 Msps) in the constructor and held forever — ~1.5 GB across a
default channel set, mostly owned by channels that were long since
tracking. The grid is now sized to the attempt's searched bin range
(~19 rows once the LO offset is known), reuses rows between attempts,
and is freed on lock and on idle; rows are f32 (each cell accumulates
only ~10 values, so no long-sum concern) with the surface statistics
still summed in f64. Channels are born in Acquisition without passing
through acquisition_start, so the grid is set up lazily on the first
process pass — caught by the synth acquire tests.

tuni2025 15 s (50 Msps): peak RSS 1.84 -> 1.12 GB; 31.0 -> 29.7 s wall.
The remaining peak is the t=0 cold start (every channel sweeps all 75
bins at once before any LO estimate exists) — staggering the cold-start
searches would shave most of it if it ever matters. CTTC fix
digit-identical; full suite green.

+++ src/util.rs
@@ -15,8 +15,8 @@ pub fn norm(v: &[Complex32]) -> f64 {
     norm_square(v).sqrt()
 }
 
-pub fn get_max_with_idx(vec: &[f64]) -> (usize, f64) {
-    let mut max = 0.0f64;
+pub fn get_max_with_idx(vec: &[f32]) -> (usize, f32) {
+    let mut max = 0.0f32;
     let mut idx = 0;
     for (i, v) in vec.iter().enumerate() {
         if *v > max {
